Purview_Get_Data_Product/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
azure-identity>=1.15.0
